Add glicko implementation.
Implemented as two methods of the Rating class - `decay` and `applyGames`. These correspond to steps 1 and 2 in the glicko paper.

`decay` varies the decay rate based on the amount of time since the last rating period, to give a fixed amount of time to decay from min deviation to max deviation.

As mentioned in the paper, the new deviation cannot get below a threshold.

Part of #14.